This commit was experimental only. Added AVL tree.
In the AVL tree, mk_avl_node each corresponding to the sched_connection was
added. AVL tree kept nodes sorted based on the value of file descriptor st-
ored. This leads to an O(lg(n)) for insertion. deletion and retrieval.
Previous implementation used a linked-list, with very fast-insertions and
deletions but an O(n) for retreival. In the current case value of n being
only 101, AVL tree implementaion ended up being just a little poorer as
compared to the just previous awesome commit.